final_project/pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastparquet>=2025.12.0",
    "google-cloud-bigquery>=3.40.0",
    "google-cloud-storage>=3.9.0",
    "jupyter>=1.1.1",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.2.1",
]
